Added homebrew bootstrapping
rake bootstrap:homebrew will install homebrew if necessary and then try
to install a number of packages. It's possible that some of these will
fail due to the need to `brew link <some other package>` first, but this
is a start.

Future: may be necessary to create scripts for some packages that do
pre/post steps?