Add install process
I wanted to make this as easy as possible to get started with and wanted
the entire process to be driven from my dotfiles repo, even though much
of the setup is actually pulled from thoughtbot's. The install process
is now:

1. Clone the repo
2. Run `install`

The `install` script will:

1. Install packages in the Brewfile.
2. Manually symlink the rcrc file so the correct settings are used on
   first `rcup` without having to pass them.
3. Run `rcup`

There is a `pre-up` hook that is run that will clone the thoughtbot
dotfiles to the proper location if that has not already been done.