Add vim-tmux-runner
My primary use for this is to run tests via vim-test. I was previously
using vim-dispatch. There are some things I like about dispatch:

1. It loads the results in the quick fix window
2. It automatically hides the quick fix window when things fail

There are something things I disliked about dispatch

1. The output was often unreadable or somehow cut off. I don't know
enough to know why that is. I often opened a split and ran tests
manually for full visibility and readability.

vim-tmux-runner is just a pane that I can send commands to. It won't
automatically hide when the tests pass. I might miss that. It will error
when I close the attached pane and try to re-run my tests (this will
most definitely annoy me until I write a vim-test strategy to overcome
this somehow), but the output will always be readable.

I also have always wanted a way to run lines from my editor in the
current debugger or whatever. So that's nice.