Don't track ssh/config
This file can contain things that, while usually not secrets, are
possibly things I don't really want to be public if I can avoid it.
There's a little value in having my config tracked in that it means I
don't have to remember the specific configuration to have macOS use the
keychain for SSH, but I can always find another way to do that (with an
init script, perhaps) that would be less succeptible to exposing
information I'd at least like to obscure.