Replace telescope with fzf-lua
I found that telescope was struggling a bit on really large projects and
I liked the idea of using native `fzf`, `bat`, `fd`, etc, as I used to
do with `Ctrl-P` in Vim.

I did quite like the appearance of telescope, but I think I can get back
to that with some more configuration if I really want to.